feat(tracing): Don't use diffy::patch since that's not working properly
