+++ Cargo.toml
@@ -19,7 +19,7 @@ mnn-sys = { version = "0.1", path = "mnn-sys", features = [] }
 thiserror = "2.0"
 error-stack.workspace = true
 oneshot = "0.1"
-tracing = { version = "0.1.40" }
+tracing = { version = "0.1.40", optional = true }
 dunce = "1.0.5"
 serde = { version = "1.0", features = ["derive"], optional = true }
 
@@ -33,8 +33,7 @@ crt_static = ["mnn-sys/crt_static"]
 # Disable mnn-threadpool to enable this
 openmp = ["mnn-sys/openmp"]
 mnn-threadpool = ["mnn-sys/mnn-threadpool"]
-# To keep compatibility with older versions newer ones automatically use it
-tracing = []
+tracing = ["dep:tracing"]
 profile = ["tracing"]
 serde = ["dep:serde"]
 
@@ -47,7 +46,6 @@ bytemuck = "1.17"
 clap = { version = "4.5", features = ["derive"] }
 divan = "0.1.14"
 tracing-subscriber = "0.3.19"
-# mnn-sync = { path = "mnn-sync" }
 
 [[bench]]
 name = "mnn-bench"
